Migrate to Copier 0.0.2
