+++ .copier-answers.yml
@@ -1,0 +1,23 @@
+# Autogenerated. Do not edit this by hand, use `copier update`.
+---
+_commit: 0.0.2
+_src_path: https://github.com/salt-extensions/salt-extension-copier
+author: Salt Core Team
+author_email: saltproject@vmware.com
+docs_url: ''
+license: apache
+loaders:
+  - auth
+  - cache
+  - module
+  - pillar
+  - returner
+  - state
+no_saltext_namespace: false
+package_name: mysql
+project_name: mysql
+salt_version: '3006'
+source_url: https://github.com/salt-extensions/saltext-mysql
+summary: Salt Extension for interacting with MySQL
+tracker_url: https://github.com/salt-extensions/saltext-mysql/issues
+url: https://github.com/salt-extensions/saltext-mysql
