+++ .github/workflows/test.yml
@@ -11,7 +11,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v1
       with:
-        python-version: 3.8
+        python-version: 3.7
     - name: Set Cache Key
       run: echo "PY=$(python --version --version | sha256sum | cut -d' ' -f1)" >> $GITHUB_ENV
     - name: Install System Deps
@@ -33,10 +33,10 @@ jobs:
     steps:
     - uses: actions/checkout@v2
 
-    - name: Set up Python 3.8 For Nox
+    - name: Set up Python 3.7 For Nox
       uses: actions/setup-python@v1
       with:
-        python-version: 3.8
+        python-version: 3.7
 
     - name: Install Nox
       run: |
@@ -63,14 +63,12 @@ jobs:
       fail-fast: false
       max-parallel: 4
       matrix:
-        include:
-          - {"salt-version": "3005.0", "python-version": "3.7"}
-          - {"salt-version": "3005.0", "python-version": "3.8"}
-          - {"salt-version": "3005.0", "python-version": "3.9"}
-          - {"salt-version": "3006.0", "python-version": "3.7"}
-          - {"salt-version": "3006.0", "python-version": "3.8"}
-          - {"salt-version": "3006.0", "python-version": "3.9"}
-          - {"salt-version": "3006.0", "python-version": "3.10"}
+        python-version:
+          - 3.6
+          - 3.7
+          - 3.8
+        salt-version:
+          - 3006
 
     steps:
     - uses: actions/checkout@v2
@@ -185,11 +183,11 @@ jobs:
       fail-fast: false
       max-parallel: 3
       matrix:
-        include:
-          - {"salt-version": "3005.0", "python-version": "3.8"}
-          - {"salt-version": "3006.0", "python-version": "3.8"}
-          - {"salt-version": "3006.0", "python-version": "3.9"}
-          - {"salt-version": "3006.0", "python-version": "3.10"}
+        python-version:
+          - 3.6
+          - 3.7
+        salt-version:
+          - 3006
 
     steps:
     - uses: actions/checkout@v2
@@ -198,15 +196,6 @@ jobs:
       uses: actions/setup-python@v1
       with:
         python-version: ${{ matrix.python-version }}
-
-    - name: Download libeay32.dll
-      run: |
-        export PY_LOC=$(which python.exe)
-        echo ${PY_LOC}
-        export PY_DIR=$(dirname ${PY_LOC})
-        echo ${PY_DIR}
-        curl https://repo.saltproject.io/windows/dependencies/64/libeay32.dll --output ${PY_DIR}/libeay32.dll
-      shell: bash
 
     - name: Install Nox
       run: |
@@ -318,10 +307,11 @@ jobs:
       fail-fast: false
       max-parallel: 3
       matrix:
-        include:
-          - {"salt-version": "3005.0", "python-version": "3.9", "source": "pypi"}
-          - {"salt-version": "3006.0", "python-version": "3.9", "source": "pypi"}
-          - {"salt-version": "3006.0", "python-version": "3.10", "source": "pypi"}
+        python-version:
+          - 3.6
+          - 3.7
+        salt-version:
+          - 3006
 
     steps:
     - uses: actions/checkout@v2
