+++ .gitignore
@@ -129,7 +129,7 @@ dmypy.json
 .pyre/
 
 # Ignore the setuptools_scm auto-generated version module
-src/saltext/saltext_mysql/version.py
+src/saltext/mysql/version.py
 
 # Ignore CI generated artifacts
 artifacts/
