+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
         name: Check CLI examples on execution modules
         entry: python .pre-commit-hooks/check-cli-examples.py
         language: system
-        files: ^src/saltext/saltext_mysql/modules/.*\.py$
+        files: ^src/saltext/mysql/modules/.*\.py$
 
   - repo: local
     hooks:
@@ -42,7 +42,7 @@ repos:
       - id: salt-rewrite
         alias: rewrite-docstrings
         name: Salt extensions docstrings auto-fixes
-        files: ^src/saltext/saltext_mysql/.*\.py$
+        files: ^src/saltext/mysql/.*\.py$
         args: [--silent]
 
   - repo: https://github.com/s0undt3ch/salt-rewrite
@@ -56,14 +56,14 @@ repos:
         args: [--silent, -E, fix_docstrings]
 
   - repo: https://github.com/asottile/pyupgrade
-    rev: v2.31.0
+    rev: v2.37.2
     hooks:
       - id: pyupgrade
         name: Rewrite Code to be Py3.6+
         args: [
           --py36-plus
         ]
-        exclude: src/saltext/saltext_mysql/version.py
+        exclude: src/saltext/mysql/version.py
 
   - repo: https://github.com/asottile/reorder_python_imports
     rev: v2.6.0
@@ -72,25 +72,25 @@ repos:
         args: [
           --py3-plus,
         ]
-        exclude: src/saltext/saltext_mysql/version.py
+        exclude: src/saltext/mysql/version.py
 
   - repo: https://github.com/psf/black
-    rev: 21.7b0
+    rev: 22.6.0
     hooks:
       - id: black
         args: [-l 100]
-        exclude: src/saltext/saltext_mysql/version.py
+        exclude: src/saltext/mysql/version.py
         additional_dependencies:
           - click<8.1.0
 
   - repo: https://github.com/asottile/blacken-docs
-    rev: v1.10.0
+    rev: v1.12.1
     hooks:
       - id: blacken-docs
         args: [--skip-errors]
-        files: ^(docs/.*\.rst|src/saltext/saltext_mysql/.*\.py)$
+        files: ^(docs/.*\.rst|src/saltext/mysql/.*\.py)$
         additional_dependencies:
-          -  black==21.7b0
+          - black==22.6.0
           - click<8.1.0
   # <---- Formatting -----------------------------------------------------------------------------
 
@@ -102,7 +102,7 @@ repos:
         alias: bandit-salt
         name: Run bandit against the code base
         args: [--silent, -lll, --skip, B701]
-        exclude: src/saltext/saltext_mysql/version.py
+        exclude: src/saltext/mysql/version.py
   - repo: https://github.com/PyCQA/bandit
     rev: "1.7.0"
     hooks:
@@ -114,17 +114,17 @@ repos:
   # <---- Security -------------------------------------------------------------------------------
 
   # ----- Code Analysis ------------------------------------------------------------------------->
-  #  - repo: https://github.com/saltstack/mirrors-nox
-  #  rev: v2021.6.12
-  #  hooks:
-  #    - id: nox
-  #      alias: lint-src
-  #      name: Lint Source Code
-  #      files: ^((setup|noxfile)|src/.*)\.py$
-  #      args:
-  #        - -e
-  #        - lint-code-pre-commit
-  #        - --
+  - repo: https://github.com/saltstack/mirrors-nox
+    rev: v2021.6.12
+    hooks:
+      - id: nox
+        alias: lint-src
+        name: Lint Source Code
+        files: ^((setup|noxfile)|src/.*)\.py$
+        args:
+          - -e
+          - lint-code-pre-commit
+          - --
 
   - repo: https://github.com/saltstack/mirrors-nox
     rev: v2021.6.12
