+++ .pre-commit-hooks/check-cli-examples.py
@@ -4,7 +4,7 @@ import re
 import sys
 
 CODE_ROOT = pathlib.Path(__file__).resolve().parent.parent
-EXECUTION_MODULES_PATH = CODE_ROOT / "src" / "saltext" / " saltext_mysql" / "modules"
+EXECUTION_MODULES_PATH = CODE_ROOT / "src" / "saltext" / " mysql" / "modules"
 
 
 def check_cli_examples(files):
