+++ requirements/base.txt
@@ -1,1 +1,1 @@
-pymysql>=1.0.2
+salt>=3006
