+++ requirements/docs.in
@@ -1,0 +1,5 @@
+sphinx
+sphinx-material-saltstack
+sphinx-prompt
+sphinxcontrib-spelling
+importlib_metadata; python_version < "3.8"
