+++ requirements/lint.in
@@ -1,0 +1,2 @@
+pylint
+saltpylint
