+++ requirements/tests.in
@@ -1,0 +1,2 @@
+pytest >= 6.1.0
+pytest-salt-factories >= 0.130.0
