+++ setup.cfg
@@ -1,15 +1,15 @@
 [metadata]
-name = saltext.saltext-mysql
+name = saltext.mysql
 description = Salt Extension for interacting with MySQL
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = Salt Core Team
 author_email = saltproject@vmware.com
 keywords = salt-extension
-url = https://github.com/saltstack/saltext-mysql
+url = https://github.com/salt-extensions/saltext-mysql
 project_urls =
-    Source=https://github.com/saltstack/saltext-mysql
-    Tracker=https://github.com/saltstack/saltext-mysql/issues
+    Source=https://github.com/salt-extensions/saltext-mysql
+    Tracker=https://github.com/salt-extensions/saltext-mysql/issues
 license = Apache Software License
 classifiers =
     Programming Language :: Python
@@ -39,7 +39,7 @@ setup_requires =
   setuptools_scm[toml]>=3.4
   setuptools-declarative-requirements
 install_requires =
-  salt>=3003
+  salt>=3006
   # Add other module install requirements above this line
 
 [options.packages.find]
@@ -51,10 +51,22 @@ exclude =
 #[options.entry_points]
 #salt.loader=
 #
+#  auth_dirs = saltext.mysql.loader:get_auth_dirs
+#
+#  cache_dirs = saltext.mysql.loader:get_cache_dirs
+#
+#  module_dirs = saltext.mysql.loader:get_module_dirs
+#
+#  pillar_dirs = saltext.mysql.loader:get_pillar_dirs
+#
+#  returner_dirs = saltext.mysql.loader:get_returner_dirs
+#
+#  state_dirs = saltext.mysql.loader:get_state_dirs
+#
 
 [options.entry_points]
 salt.loader=
-  saltext.saltext-mysql = saltext.saltext_mysql
+  saltext.mysql = saltext.mysql
 
 [requirements-files]
 install_requires = requirements/base.txt
