--- src/saltext/saltext_mysql/loader.py
@@ -1,5 +1,0 @@
-"""
-Define the required entry-points functions in order for Salt to know
-what and from where it should load this extension's loaders
-"""
-from . import PACKAGE_ROOT  # pylint: disable=unused-import
