+++ tests/conftest.py
@@ -1,17 +1,30 @@
+import logging
 import os
 
 import pytest
-from saltext.saltext_mysql import PACKAGE_ROOT
+from saltext.mysql import PACKAGE_ROOT
 from saltfactories.utils import random_string
+
+
+# Reset the root logger to its default level(because salt changed it)
+logging.root.setLevel(logging.WARNING)
+
+
+# This swallows all logging to stdout.
+# To show select logs, set --log-cli-level=<level>
+for handler in logging.root.handlers[:]:
+    logging.root.removeHandler(handler)
+    handler.close()
 
 
 @pytest.fixture(scope="session")
 def salt_factories_config():
     """
-    Return a dictionary with the keyworkd arguments for FactoriesManager
+    Return a dictionary with the keyword arguments for FactoriesManager
     """
     return {
         "code_dir": str(PACKAGE_ROOT),
+        "inject_sitecustomize": "COVERAGE_PROCESS_START" in os.environ,
         "start_timeout": 120 if os.environ.get("CI") else 60,
     }
